feat: add test and separate column's tests
